+++ react/__mocks__/vtex.order-manager/OrderForm.ts
@@ -1,0 +1,5 @@
+export const useOrderForm = () => {
+  return {
+    orderForm: { id: 'mockedOrderFormId' },
+  }
+}